// fe/src/lib/api/generated/model/counselingContextSharedSummariesResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { ContextSharedSummaryResponse } from './contextSharedSummaryResponse';

export interface CounselingContextSharedSummariesResponse {
  items: ContextSharedSummaryResponse[];
  limit: number;
}