// fe/src/lib/api/generated/model/counselingContextAnchorType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CounselingContextAnchorType = typeof CounselingContextAnchorType[keyof typeof CounselingContextAnchorType];


export const CounselingContextAnchorType = {
  APPOINTMENT: 'APPOINTMENT',
  ROUTINE_INTERVIEW: 'ROUTINE_INTERVIEW',
} as const;