// fe/src/lib/api/generated/model/counselingContextInventoryStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CounselingContextInventoryStatus = typeof CounselingContextInventoryStatus[keyof typeof CounselingContextInventoryStatus];


export const CounselingContextInventoryStatus = {
  MISSING: 'MISSING',
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
} as const;